broke out Folder Images settings tab
